Move lib/devise/* to lib/devise/strategies/* (#27638)
