+++ config/application.rb
@@ -39,8 +39,8 @@ require_relative '../lib/mastodon/snowflake'
 require_relative '../lib/mastodon/version'
 require_relative '../lib/mastodon/rack_middleware'
 require_relative '../lib/public_file_server_middleware'
-require_relative '../lib/devise/two_factor_ldap_authenticatable'
-require_relative '../lib/devise/two_factor_pam_authenticatable'
+require_relative '../lib/devise/strategies/two_factor_ldap_authenticatable'
+require_relative '../lib/devise/strategies/two_factor_pam_authenticatable'
 require_relative '../lib/chewy/settings_extensions'
 require_relative '../lib/chewy/index_extensions'
 require_relative '../lib/chewy/strategy/mastodon'
