+++ lib/devise/strategies/two_factor_ldap_authenticatable.rb
